+++ RELEASE_NOTES.md
@@ -1,6 +1,12 @@
 Release Notes
 =============
 
+## Next Release
+
+- `Row.columns` returns the names of columns in the row.
+- `Row.databaseValues` returns the database values in the row.
+
+
 ## 0.11.0
 
 Released September 4, 2015